+++ gemini/collapse.gemini.js
@@ -1,6 +1,6 @@
-import Collapse from '../src/collapse';
+import Collapse from '../src/collapse/fantasy';
 import GeminiBox from '../gemini-utils/gemini-box/gemini-box';
-import Paragraph from '../src/paragraph';
+import Paragraph from '../src/paragraph/fantasy';
 import { LOREM_IPSUM } from '../src/vars';
 
 const NAME = 'collapse';